+++ packages/code-editor/src/plugins/prettier.tsx
@@ -29,7 +29,11 @@ const formatCode = async (code: string) => {
 };
 
 const usePrettier = () => {
-  const registerPrettier = async (_editor: EditorInstance, monaco: Monaco, options?: { theme?: string }) => {
+  const registerPrettier = async (
+    _editor: EditorInstance,
+    monaco: Monaco,
+    options: { theme?: string; typeScriptNoValidation?: boolean }
+  ) => {
     monaco.languages.registerDocumentFormattingEditProvider(['javascript', 'typescript'], {
       async provideDocumentFormattingEdits(model) {
         return [
@@ -43,7 +47,7 @@ const usePrettier = () => {
 
     monaco.languages.typescript.typescriptDefaults.setDiagnosticsOptions({
       noSemanticValidation: true,
-      noSyntaxValidation: true,
+      noSyntaxValidation: options?.typeScriptNoValidation ?? true,
     });
 
     if (options?.theme) {